+++ spec/classes/lsys_nginx_spec.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe 'lsys_nginx' do
+  on_supported_os.each do |os, os_facts|
+    context "on #{os}" do
+      let(:facts) { os_facts }
+
+      it { is_expected.to compile.with_all_deps }
+
+      case os
+      when 'redhat-8-x86_64', 'centos-8-x86_64'
+        it {
+          is_expected.to contain_package('nginx')
+            .with_ensure(%r{\.el8})
+            .with_provider('dnf')
+            .with_install_options(['--repo', 'nginx-release'])
+        }
+      end
+    end
+  end
+end